\name{Disaggregate}
\alias{Disaggregate}
\title{Disaggregates an rts object with a higher frequency rts object}
\usage{
  Disaggregate(agg, ind)
}
\arguments{
  \item{agg}{a rts object of a lower frequency}

  \item{ind}{a rts object of a higher frequency indicator}
}
\value{
  An rts with the same freq as ind and the same id as agg
}
\description{
  Disaggregates an rts object with a higher frequency rts
  object
}

