\name{DocRemarkRts}
\alias{DocRemarkRts}
\title{adds a remark to all doc objects of an rts object}
\usage{
  DocRemarkRts(x, text, start = NULL, end = NULL)
}
\arguments{
  \item{x}{rts object}

  \item{text}{char string}

  \item{start}{numeric}

  \item{end}{numeric}
}
\value{
  rts object
}
\description{
  adds a remark to all doc objects of an rts object
}

