\name{IndexBetweenPriority}
\alias{IndexBetweenPriority}
\title{an index of the between.prority in a rts.list}
\usage{
  IndexBetweenPriority(rts.list)
}
\arguments{
  \item{rts.list}{rts.list}
}
\value{
  an index of the frequencies in a rts.list
}
\description{
  an index of the between.prority in a rts.list
}

