\name{IndexWithinPriority}
\alias{IndexWithinPriority}
\title{an index of the within.priority in a rts.list}
\usage{
  IndexWithinPriority(rts.list)
}
\arguments{
  \item{rts.list}{rts.list}
}
\value{
  an index of the frequencies in a rts.list
}
\description{
  an index of the within.priority in a rts.list
}

