\name{plot.mrts}
\alias{plot.mrts}
\title{plot method for an mrts object}
\usage{
  \method{plot}{mrts} (x, ...)
}
\arguments{
  \item{x}{an mrts object}

  \item{...}{unused, for compatibility with the generic}
}
\value{
  plots the object as a side effect
}
\description{
  plot method for an mrts object
}

