\name{plot.rts}
\alias{plot.rts}
\title{plot method for an rts object}
\usage{
  \method{plot}{rts} (x, ...)
}
\arguments{
  \item{x}{an rts object}

  \item{...}{unused, for compatibility with the generic}
}
\value{
  plots the object as a side effect
}
\description{
  plot method for an rts object
}

