\name{SortMergeRts}
\alias{SortMergeRts}
\title{Sort and merges mts objects of a given frequency according their priority}
\usage{
  SortMergeRts(x)
}
\arguments{
  \item{x}{a list with rts objects}
}
\value{
  An rts with the same freq as ind and the same id as agg
}
\description{
  Sort and merges mts objects of a given frequency
  according their priority
}

