\name{window.rts}
\alias{window.rts}
\title{window method for rts class}
\usage{
  \method{window}{rts} (x, start = NULL, end = NULL, ...)
}
\arguments{
  \item{x}{rts object}

  \item{start}{start of rts}

  \item{end}{end of rts}

  \item{...}{unused, for compatibility with the generic}
}
\value{
  rts object
}
\description{
  window method for rts class
}

